Rough working draft of fragment randomisation.

+++ SubnauticaRandomiser/RandomiserObjects/Biome.cs
@@ -6,14 +6,14 @@ namespace SubnauticaRandomiser.RandomiserObjects
         public readonly int CreatureSlots;
         public readonly int MediumSlots;
         public readonly int SmallSlots;
-        public readonly int? FragmentRate;
+        public readonly float? FragmentRate;
         public readonly string Name;
         public readonly EBiomeType BiomeType;
 
         public int AverageDepth { get { return BiomeType.GetAccessibleDepth(); } }
         public int Used = 0;
 
-        public Biome(string name, EBiomeType biomeType, int creatureSlots, int mediumSlots, int smallSlots = -1, int? fragmentRate = null)
+        public Biome(string name, EBiomeType biomeType, int creatureSlots, int mediumSlots, int smallSlots = -1, float? fragmentRate = null)
         {
             Name = name;
             BiomeType = biomeType;
